Add row level validation
Add the concept of row-level validation and a basic RowLengthValidator
as a proof of concept / first use. This change allows for validators
that make use of the row as a whole rather than individual fields. The
RowLengthValidator demonstrates this by simply checking the length of
the row (a dict returned by csv.DictReader); other uses may include
interdependent fields (e.g. A and B are both present or both absent).